Final Commit

+++ webadmin/invoice.php
@@ -4,6 +4,8 @@ session_start();
 include("./inc/constant.inc.php");
 include("./inc/connection.inc.php");
 require_once("./inc/smtp/class.phpmailer.php");
+require_once("./inc/phpqrcode/qrlib.php");
+require('vendor/autoload.php');
 if (isset($_GET['id']) && $_GET['id']!="") {
     $invoice_id=get_safe_value($_GET['id']);
         if(isset($_GET['send_email']) && $_GET['send_email']=='true'){
@@ -12,11 +14,13 @@ if (isset($_GET['id']) && $_GET['id']!="") {
     $_SESSION['PERMISSION_ERROR']=1;
     redirect("index.php");
 }
-require('vendor/autoload.php');
 $invoice_id=$_GET['id'];
-$sql="select total_amount from payments where id='$invoice_id'";
+$sql="select total_amount from payments where md5(id)='$invoice_id'";
 $res=mysqli_query($con,$sql);
 $html="";
+$filepath = 'qrcode.png';
+QRcode::png(QR_CODE_ADDRESS."invoice.php?id=".$invoice_id, $filepath);
+
 if(mysqli_num_rows($res)>0){
     $row=mysqli_fetch_assoc($res);
     $total_amount=$row['total_amount'];
@@ -24,24 +28,24 @@ if(mysqli_num_rows($res)>0){
     $html.='
         <tr>
             <td align="center">                    
-                <img width="150" src="./img/logo.jpg" width="100" height="100" />
+                <img width="150" src="'.LOGO.'" width="100" height="100" />
             </td>
             <td  align="center" colspan="2">
-                <strong><span style="font-size:25px">Barisal Engineering College Hall</span></strong>
+                <strong><span style="font-size:25px">'.HALL_NAME.'</span></strong>
                 <br>
                 Durgapur, Barisal
                 <br>
-                Tel: +00 000 000 0000 | Email: becians2017@gmail.com
+                Tel: '.TEL.' | Email: '.EMAIL.'
                 <br>
-                http://www.barisal-eng.edu.bd/
+                '.WEBSITE.'
             </td>
         </tr>';
         $html.='
             <tr><td colspan="3"><hr></td></tr>
             <tr width="100%">
             </tr>';
-            $sql="select users.id,users.name,users.batch,users.dept_id,users.roll,depts.id,depts.name as dept_name, payments.id,payments.user_id,payments.created_at from users,depts, payments where users.id=payments.user_id and users.dept_id=depts.id and payments.id='$invoice_id'";
-            $res=mysqli_query($con,$sql);
+            $user_sql="select users.id,users.name,users.batch,users.dept_id,users.class_roll,depts.id,depts.name as dept_name, payments.id as payment_id,payments.user_id,payments.created_at from users,depts, payments where users.id=payments.user_id and users.dept_id=depts.id and md5(payments.id)='$invoice_id'";
+            $res=mysqli_query($con,$user_sql);
             if(mysqli_num_rows($res)>0){
                 while($row=mysqli_fetch_assoc($res)){
             $html.='
@@ -51,17 +55,17 @@ if(mysqli_num_rows($res)>0){
                     </td>
                     <td rowspan="3" align="center" >   
                         <span style="font-size:20px; background-color: #b7b4b4">Student\'s Copy</span><br>
-                        <img width="150" src="https://chart.googleapis.com/chart?cht=qr&chs=500x500&chl='.FRONT_SITE_PATH.'/webadmin/invoice.php?id='.$invoice_id.'&choe=UTF-8"  width="100" height="100" />
+                        <img width="150" src="'.$filepath.'" width="100" height="100" />
                         <br>Scan QR to verify Payment
                     </td>
                     <td style="text-align:right;border: 1px solid black;border-collapse: collapse;" >
-                        Invoice No: #'.$invoice_id.'
+                        Invoice No: #'.$row["payment_id"].'
                     </td>
                 </tr>';
                 $html.='
                     <tr width="100%">
                         <td  style="border: 1px solid black;border-collapse: collapse;padding-left: 20px;">
-                        Roll : '.$row["roll"].'
+                        Roll : '.$row["class_roll"].'
                         </td>
                         <td style="text-align:right;border: 1px solid black;border-collapse: collapse;" >
                             Created: #'.date("d M Y h:i A",$row["created_at"]).'
@@ -85,12 +89,12 @@ if(mysqli_num_rows($res)>0){
                     </td>
                 </tr>';
                 $html.='<tr>
-                        <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">Id</td>
+                        <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">SL No</td>
                         <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">Fee details</td>
                         <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">Amount</td>
                     </tr>';
-                    $sql="select payments.*,monthly_payment_details.*,month.id,month.name from payments,monthly_payment_details,month where month.id=monthly_payment_details.month_id and monthly_payment_details.payment_id=payments.id and payments.id='$invoice_id'";
-                    $res=mysqli_query($con,$sql);
+                    $payment_sql="select payments.*,monthly_payment_details.*,month.id,month.name from payments,monthly_payment_details,month where month.id=monthly_payment_details.month_id and monthly_payment_details.payment_id=payments.id and md5(payments.id)='$invoice_id'";
+                    $res=mysqli_query($con,$payment_sql);
                     if(mysqli_num_rows($res)>0){
                     $i=1;
                         while($row=mysqli_fetch_assoc($res)){
@@ -115,15 +119,15 @@ if(mysqli_num_rows($res)>0){
                     </td>
                 </tr>';
                 $html.='<tr>
-                        <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">Id</td>
+                        <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">SL No</td>
                         <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">Fee details</td>
                         <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">Amount</td>
                     </tr>';
-                $sql="SELECT `fees`.* , fee_details.*,payments.* from payments,fees,fee_details WHERE fees.id=fee_details.fee_id and payments.id=fee_details.payment_id and payments.id='$invoice_id'";
-                $res=mysqli_query($con,$sql);
-                if(mysqli_num_rows($res)>0){
+                $fee_sql="SELECT `fees`.* , fee_details.*,payments.* from payments,fees,fee_details WHERE fees.id=fee_details.fee_id and payments.id=fee_details.payment_id and md5(payments.id)='$invoice_id'";
+                $fee_res=mysqli_query($con,$fee_sql);
+                if(mysqli_num_rows($fee_res)>0){
                     $i=1;
-                    while($row=mysqli_fetch_assoc($res)){
+                    while($row=mysqli_fetch_assoc($fee_res)){
                         $html.='
                             <tr>
                                 <td style="border: 1px solid black;border-collapse: collapse;text-align:center;">'.$i.'</td>
@@ -158,9 +162,7 @@ if(mysqli_num_rows($res)>0){
                 </tr>';
                 $html.='
                     <tr width="100%">
-                    </tr>';
-                    $sql="select users.id,users.name,users.batch,users.dept_id,users.roll,depts.id,depts.name as dept_name, payments.id,payments.user_id,payments.created_at from users,depts, payments where users.id=payments.user_id and users.dept_id=depts.id and payments.id='$invoice_id'";
-                    $res=mysqli_query($con,$sql);
+                    </tr>';$res=mysqli_query($con,$user_sql);
                     if(mysqli_num_rows($res)>0){
                         while($row=mysqli_fetch_assoc($res)){
                     $html.='
@@ -170,10 +172,10 @@ if(mysqli_num_rows($res)>0){
                             </td>
                             <td rowspan="3" align="center" >   
                                 <span style="font-size:10px; background-color: #b7b4b4">Office Copy</span><br>
-                                <img width="150" src="https://chart.googleapis.com/chart?cht=qr&chs=500x500&chl='.FRONT_SITE_PATH.'/invoice.php?id='.$invoice_id.'&choe=UTF-8"  width="70" height="70" />
+                                <img width="150" src="'.$filepath.'"  width="70" height="70" />
                             </td>
                             <td style="text-align:right;border: 1px solid black;border-collapse: collapse;" >
-                                Invoice No: #'.$invoice_id.'
+                                Invoice No: #'.$row["payment_id"].'
                             </td>
                         </tr>';
                         $html.='
@@ -203,15 +205,14 @@ if(mysqli_num_rows($res)>0){
                     </td>
                 </tr>';
                 $html.='<tr>
-                        <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">Id</td>
+                        <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">SL No</td>
                         <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">Fee details</td>
                         <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">Amount</td>
                     </tr>';
-                    $sql="select payments.*,monthly_payment_details.*,month.id,month.name from payments,monthly_payment_details,month where month.id=monthly_payment_details.month_id and monthly_payment_details.payment_id=payments.id and payments.id='$invoice_id'";
-                    $res=mysqli_query($con,$sql);
-                    if(mysqli_num_rows($res)>0){
+                    $payment_res=mysqli_query($con,$payment_sql);
+                    if(mysqli_num_rows($payment_res)>0){
                     $i=1;
-                    while($row=mysqli_fetch_assoc($res)){
+                    while($row=mysqli_fetch_assoc($payment_res)){
                         $html.='<tr>
                         <td style="border: 1px solid black;border-collapse: collapse;text-align:center;">'.$i.'</td>
                         <td style="border: 1px solid black;border-collapse: collapse;text-align:center;">'.$row['name']." - ".date("y",time()).'</td>
@@ -233,12 +234,11 @@ if(mysqli_num_rows($res)>0){
                     </td>
                 </tr>';
                 $html.='<tr>
-                        <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">Id</td>
+                        <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">SL No</td>
                         <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">Fee details</td>
                         <td style="border: 1px solid black;border-collapse: collapse;background-color: #b7b4b4;text-align:center;">Amount</td>
                     </tr>';
-                $sql="SELECT `fees`.* , fee_details.*,payments.* from payments,fees,fee_details WHERE fees.id=fee_details.fee_id and payments.id=fee_details.payment_id and payments.id='$invoice_id'";
-                $res=mysqli_query($con,$sql);
+                $res=mysqli_query($con,$fee_sql);
                 if(mysqli_num_rows($res)>0){
                     $i=1;
                     while($row=mysqli_fetch_assoc($res)){
@@ -266,6 +266,7 @@ if(mysqli_num_rows($res)>0){
 }else{
     $html.="No data found";
 }
+// echo $html;
 $mpdf=new \Mpdf\Mpdf([
     'tempDir' => __DIR__ . '/custom/temp/dir/path',
     'default_font_size' => 12,
@@ -275,15 +276,17 @@ $mpdf=new \Mpdf\Mpdf([
 	'margin_top' => 2,
 	'margin_bottom' => 10,
 ]);
-$mpdf->SetCreator('Dhrubo');
-$mpdf->SetAuthor('Dhrubo');
+// $mpdf->SetCreator('Dhrubo');
+// $mpdf->SetAuthor('Dhrubo');
+$mpdf->SetProtection(array('print'), '', 'MyPassword');
 $mpdf->SetTitle('Invoice of Barisal Engineering College Hall Payment');
 $mpdf->SetFooter('Developed By The Web divers');
 $mpdf->watermarkImageAlpha = 0.1;
 $mpdf->WriteHTML($html);
-$file=time().'.pdf';
-$mpdf->output($file,'I');
-// $mpdf->output($file,'D');
-$mpdf->output('media/'.$file,'F');
-send_email("dhruborajroy3@gmail.com","Invoice","Invoice of Payment ".$file,'media/'.$file);
-unlink($file);
+$file = time() . '.pdf';
+$mpdf->output($file, 'F'); 
+$mpdf->output($file, 'I'); 
+send_email("dhruborajroy3@gmail.com", "Invoice", "Invoice of Payment", $file);
+if (file_exists($file)) {
+    unlink($file); 
+}
